UPSTREAM: ext4: use private version of page_zero_new_buffers() for data=journal mode
commit b90197b655185a11640cce3a0a0bc5d8291b8ad2 upstream.

If there is a error while copying data from userspace into the page
cache during a write(2) system call, in data=journal mode, in
ext4_journalled_write_end() were using page_zero_new_buffers() from
fs/buffer.c.  Unfortunately, this sets the buffer dirty flag, which is
no good if journalling is enabled.  This is a long-standing bug that
goes back for years and years in ext3, but a combination of (a)
data=journal not being very common, (b) in many case it only results
in a warning message. and (c) only very rarely causes the kernel hang,
means that we only really noticed this as a problem when commit
998ef75ddb caused this failure to happen frequently enough to cause
generic/208 to fail when run in data=journal mode.

The fix is to have our own version of this function that doesn't call
mark_dirty_buffer(), since we will end up calling
ext4_handle_dirty_metadata() on the buffer head(s) in questions very
shortly afterwards in ext4_journalled_write_end().

Thanks to Dave Hansen and Linus Torvalds for helping to identify the
root cause of the problem.

BUG=chromium:766786
TEST=build/boot on caroline, xfstest

Change-Id: I9309bff07aded1e0dc2c7272461aa422ad2cb71f
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Reviewed-by: Jan Kara <jack@suse.com>
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>
(cherry picked from commit dabb07246427d46ece1a68a02e0e67fbb46e53fa
 from 3.16 stable)
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/689215
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Signed-off-by: Harsh Shandilya <harsh@prjkt.io>

Signed-off-by: AyushR1 <ayushrathore12501@gmail.com>